Loops for printing are now 64-bit unsigned ints since the number of nodes is much larger 65,000.  It does compile but I have yet to test this fix.

+++ src/c/bpopsim/cSubpopulation.cc
@@ -71,7 +71,7 @@ void cSubpopulation::NewCreateDescendant(gsl_rng * randomgenerator,
 void cSubpopulation::Transfer(long double success_prob, 
                               gsl_rng * randomgenerator) {
 		
-	int random_gsl_int = gsl_ran_binomial(randomgenerator, 
+	u_int64_t random_gsl_int = gsl_ran_binomial(randomgenerator, 
                                           success_prob, 
                                           u_int64_t(GetNumber()));
